數字轉換問題

+++ Repository/ConversationRepository.cs
@@ -156,7 +156,7 @@ namespace OpenLineBot.Repository {
             DocumentSnapshot docShot = await _db.Collection (userId).Document (bookDate).GetSnapshotAsync ();
             Dictionary<string, object> record = new Dictionary<string, object> { { "Id", Guid.NewGuid().ToString("N") },
                 { "Name", item },
-                { "Money", Convert.ToInt32 (money) }
+                { "Money", Convert.ToDecimal(money) }
             };
             if (docShot.Exists) {
                 if (docShot.GetValue<List<Dictionary<string, object>>> ("list").Count > 0) {
